requirements.txt
africastalking==1.2.7
antiorm==1.2.1
argcomplete==3.1.6
asgiref==3.7.2
boto3==1.29.2
botocore==1.32.2
certifi==2023.7.22
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
contextlib2==21.6.0
db==0.1.1
db-sqlite3==0.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
durationpy==0.5
environ==1.0
gunicorn==21.2.0
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
packaging==23.2
placebo==0.9.0
pysqlite-binary==0.5.1.3380300
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
schema==0.7.5
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.2
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.1
zappa==0.58.0
